refactor: consolidate UserSummary into shared common.py schema
