+++ app/schemas/image.py
@@ -9,25 +9,7 @@ from pydantic import BaseModel, Field, computed_field, field_validator
 
 from app.config import TagType, settings
 from app.models.image import ImageBase
-
-
-class UserSummary(BaseModel):
-    """Minimal user info for embedding"""
-
-    user_id: int
-    username: str
-    avatar: str | None = None  # Avatar filename from database
-
-    # Allow Pydantic to read from SQLAlchemy model attributes (not just dicts)
-    model_config = {"from_attributes": True}
-
-    @computed_field  # type: ignore[prop-decorator]
-    @property
-    def avatar_url(self) -> str | None:
-        """Generate avatar URL from avatar field"""
-        if self.avatar:
-            return f"{settings.IMAGE_BASE_URL}/storage/avatars/{self.avatar}"
-        return None
+from app.schemas.common import UserSummary
 
 
 class TagSummary(BaseModel):
